backend/requirements.txt
Django==5.2.10
gunicorn==23.0.0
psycopg2-binary==2.9.11
markdown==3.10
django-filter==25.2
django-cors-headers==4.9.0
pillow==12.1.0
python-dateutil==2.9.0
django-dbbackup==5.1.0
django-ninja==1.5.2
python-decouple==3.8
django-q2==1.9.0
django-jazzmin==3.0.1
django-import-export==4.3.14
arrow==1.4.0
pytz==2025.2
ms-active-directory==1.14.1
pytest==9.0.2
pytest-django==4.11.1
django-model-utils==5.0.0
pytest-mock==3.15.1
django-allauth==65.13.1